RF: Move mark_sensitive into dlad.py module
